implementasi navigate to chat room from book card
