+++ app/src/components/HomePage/Pagination.jsx
@@ -1,0 +1,102 @@
+import { ChevronLeft, ChevronRight } from "lucide-react";
+
+// buat pagination
+export const Pagination = ({ currentPage, totalPages, onPageChange }) => {
+  const renderPageNumbers = () => {
+    const pages = [];
+    const maxPagesToShow = 5;
+
+    let startPage = Math.max(1, currentPage - Math.floor(maxPagesToShow / 2));
+    let endPage = Math.min(totalPages, startPage + maxPagesToShow - 1);
+
+    if (endPage - startPage + 1 < maxPagesToShow) {
+      startPage = Math.max(1, endPage - maxPagesToShow + 1);
+    }
+
+    if (startPage > 1) {
+      pages.push(
+        <button
+          key={1}
+          onClick={() => onPageChange(1)}
+          className="px-3 py-1 mx-1 rounded hover:bg-gray-200"
+        >
+          1
+        </button>
+      );
+      if (startPage > 2) {
+        pages.push(<span key="ellipsis1">...</span>);
+      }
+    }
+
+    for (let i = startPage; i <= endPage; i++) {
+      pages.push(
+        <button
+          key={i}
+          onClick={() => onPageChange(i)}
+          className={`px-3 py-1 mx-1 rounded ${
+            currentPage === i ? "bg-[#00A8FF] text-white" : "hover:bg-gray-200"
+          }`}
+        >
+          {i}
+        </button>
+      );
+    }
+
+    if (endPage < totalPages) {
+      if (endPage < totalPages - 1) {
+        pages.push(<span key="ellipsis2">...</span>);
+      }
+      pages.push(
+        <button
+          key={totalPages}
+          onClick={() => onPageChange(totalPages)}
+          className="px-3 py-1 mx-1 rounded hover:bg-gray-200"
+        >
+          {totalPages}
+        </button>
+      );
+    }
+
+    return pages;
+  };
+
+  return (
+    <div className="flex justify-center items-center my-8">
+      <button
+        onClick={() => onPageChange(currentPage - 1)}
+        disabled={currentPage === 1}
+        className={`flex items-center p-2 rounded ${
+          currentPage === 1
+            ? "text-gray-400 cursor-not-allowed"
+            : "hover:bg-gray-200 text-gray-700"
+        }`}
+      >
+        <ChevronLeft size={16} />
+        <span className="ml-1">Prev</span>
+      </button>
+
+      <div className="hidden md:flex items-center mx-4">
+        {renderPageNumbers()}
+      </div>
+
+      <div className="md:hidden mx-4">
+        <span>
+          Page {currentPage} of {totalPages}
+        </span>
+      </div>
+
+      <button
+        onClick={() => onPageChange(currentPage + 1)}
+        disabled={currentPage === totalPages}
+        className={`flex items-center p-2 rounded ${
+          currentPage === totalPages
+            ? "text-gray-400 cursor-not-allowed"
+            : "hover:bg-gray-200 text-gray-700"
+        }`}
+      >
+        <span className="mr-1">Next</span>
+        <ChevronRight size={16} />
+      </button>
+    </div>
+  );
+};
